+++ zarrs_plugin/src/plugin.rs
@@ -1,25 +1,34 @@
 use crate::PluginCreateError;
 
 /// A plugin.
-pub struct Plugin<TPlugin, TInput, TError = PluginCreateError> {
+pub struct Plugin<TPlugin, TInput, TError = PluginCreateError, TMatch = str>
+where
+    TMatch: ?Sized,
+{
     /// Tests if the name is a match for this plugin.
-    match_name_fn: fn(name: &str) -> bool,
+    match_name_fn: fn(name: &TMatch) -> bool,
     /// Create an implementation of this plugin from metadata.
     create_fn: fn(input: &TInput) -> Result<TPlugin, TError>,
 }
 
 /// A plugin (two parameters).
-pub struct Plugin2<TPlugin, TInput1, TInput2, TError = PluginCreateError> {
+pub struct Plugin2<TPlugin, TInput1, TInput2, TError = PluginCreateError, TMatch = str>
+where
+    TMatch: ?Sized,
+{
     /// Tests if the name is a match for this plugin.
-    match_name_fn: fn(name: &str) -> bool,
+    match_name_fn: fn(name: &TMatch) -> bool,
     /// Create an implementation of this plugin from metadata.
     create_fn: fn(input1: &TInput1, input2: &TInput2) -> Result<TPlugin, TError>,
 }
 
-impl<TPlugin, TInput, TError> Plugin<TPlugin, TInput, TError> {
+impl<TPlugin, TInput, TError, TMatch> Plugin<TPlugin, TInput, TError, TMatch>
+where
+    TMatch: ?Sized,
+{
     /// Create a new plugin for registration.
     pub const fn new(
-        match_name_fn: fn(name: &str) -> bool,
+        match_name_fn: fn(name: &TMatch) -> bool,
         create_fn: fn(inputs: &TInput) -> Result<TPlugin, TError>,
     ) -> Self {
         Self {
@@ -39,15 +48,18 @@ impl<TPlugin, TInput, TError> Plugin<TPlugin, TInput, TError> {
 
     /// Returns true if this plugin is associated with `name`.
     #[must_use]
-    pub fn match_name(&self, name: &str) -> bool {
+    pub fn match_name(&self, name: &TMatch) -> bool {
         (self.match_name_fn)(name)
     }
 }
 
-impl<TPlugin, TInput1, TInput2, TError> Plugin2<TPlugin, TInput1, TInput2, TError> {
+impl<TPlugin, TInput1, TInput2, TError, TMatch> Plugin2<TPlugin, TInput1, TInput2, TError, TMatch>
+where
+    TMatch: ?Sized,
+{
     /// Create a new plugin for registration.
     pub const fn new(
-        match_name_fn: fn(name: &str) -> bool,
+        match_name_fn: fn(name: &TMatch) -> bool,
         create_fn: fn(input1: &TInput1, input2: &TInput2) -> Result<TPlugin, TError>,
     ) -> Self {
         Self {
@@ -70,7 +82,7 @@ impl<TPlugin, TInput1, TInput2, TError> Plugin2<TPlugin, TInput1, TInput2, TErro
 
     /// Returns true if this plugin is associated with `name`.
     #[must_use]
-    pub fn match_name(&self, name: &str) -> bool {
+    pub fn match_name(&self, name: &TMatch) -> bool {
         (self.match_name_fn)(name)
     }
 }