+++ zarrs_plugin/src/runtime_plugin.rs
@@ -22,26 +22,28 @@ use crate::{MaybeSend, MaybeSync, PluginCreateError};
 /// );
 /// ```
 #[allow(clippy::type_complexity)]
-pub struct RuntimePlugin<TPlugin, TInput, TError = PluginCreateError>
+pub struct RuntimePlugin<TPlugin, TInput, TError = PluginCreateError, TMatch = str>
 where
     TPlugin: MaybeSend + MaybeSync + 'static,
     TInput: ?Sized + 'static,
+    TMatch: ?Sized,
 {
     /// Tests if the name is a match for this plugin.
-    match_name_fn: Box<dyn Fn(&str) -> bool + Send + Sync + 'static>,
+    match_name_fn: Box<dyn Fn(&TMatch) -> bool + Send + Sync + 'static>,
     /// Create an implementation of this plugin from input.
     create_fn: Box<dyn Fn(&TInput) -> Result<TPlugin, TError> + Send + Sync + 'static>,
 }
 
-impl<TPlugin, TInput, TError> RuntimePlugin<TPlugin, TInput, TError>
+impl<TPlugin, TInput, TError, TMatch> RuntimePlugin<TPlugin, TInput, TError, TMatch>
 where
     TPlugin: MaybeSend + MaybeSync + 'static,
     TInput: ?Sized + 'static,
+    TMatch: ?Sized,
 {
     /// Create a new runtime plugin for registration.
     pub fn new<M, C>(match_name_fn: M, create_fn: C) -> Self
     where
-        M: Fn(&str) -> bool + Send + Sync + 'static,
+        M: Fn(&TMatch) -> bool + Send + Sync + 'static,
         C: Fn(&TInput) -> Result<TPlugin, TError> + Send + Sync + 'static,
     {
         Self {
@@ -60,7 +62,7 @@ where
 
     /// Returns true if this plugin is associated with `name`.
     #[must_use]
-    pub fn match_name(&self, name: &str) -> bool {
+    pub fn match_name(&self, name: &TMatch) -> bool {
         (self.match_name_fn)(name)
     }
 }
@@ -69,28 +71,31 @@ where
 ///
 /// This is the runtime equivalent of [`Plugin2`](crate::Plugin2).
 #[allow(clippy::type_complexity)]
-pub struct RuntimePlugin2<TPlugin, TInput1, TInput2, TError = PluginCreateError>
+pub struct RuntimePlugin2<TPlugin, TInput1, TInput2, TError = PluginCreateError, TMatch = str>
 where
     TPlugin: MaybeSend + MaybeSync + 'static,
     TInput1: ?Sized + 'static,
     TInput2: ?Sized + 'static,
+    TMatch: ?Sized,
 {
     /// Tests if the name is a match for this plugin.
-    match_name_fn: Box<dyn Fn(&str) -> bool + Send + Sync + 'static>,
+    match_name_fn: Box<dyn Fn(&TMatch) -> bool + Send + Sync + 'static>,
     /// Create an implementation of this plugin from input.
     create_fn: Box<dyn Fn(&TInput1, &TInput2) -> Result<TPlugin, TError> + Send + Sync + 'static>,
 }
 
-impl<TPlugin, TInput1, TInput2, TError> RuntimePlugin2<TPlugin, TInput1, TInput2, TError>
+impl<TPlugin, TInput1, TInput2, TError, TMatch>
+    RuntimePlugin2<TPlugin, TInput1, TInput2, TError, TMatch>
 where
     TPlugin: MaybeSend + MaybeSync + 'static,
     TInput1: ?Sized + 'static,
     TInput2: ?Sized + 'static,
+    TMatch: ?Sized,
 {
     /// Create a new runtime plugin for registration.
     pub fn new<M, C>(match_name_fn: M, create_fn: C) -> Self
     where
-        M: Fn(&str) -> bool + Send + Sync + 'static,
+        M: Fn(&TMatch) -> bool + Send + Sync + 'static,
         C: Fn(&TInput1, &TInput2) -> Result<TPlugin, TError> + Send + Sync + 'static,
     {
         Self {
@@ -109,7 +114,7 @@ where
 
     /// Returns true if this plugin is associated with `name`.
     #[must_use]
-    pub fn match_name(&self, name: &str) -> bool {
+    pub fn match_name(&self, name: &TMatch) -> bool {
         (self.match_name_fn)(name)
     }
 }
@@ -121,12 +126,12 @@ where
 mod wasm_impls {
     use super::{RuntimePlugin, RuntimePlugin2};
 
-    unsafe impl<TPlugin: 'static, TInput: ?Sized + 'static, TError: 'static> Send
-        for RuntimePlugin<TPlugin, TInput, TError>
+    unsafe impl<TPlugin: 'static, TInput: ?Sized + 'static, TError: 'static, TMatch: ?Sized> Send
+        for RuntimePlugin<TPlugin, TInput, TError, TMatch>
     {
     }
-    unsafe impl<TPlugin: 'static, TInput: ?Sized + 'static, TError: 'static> Sync
-        for RuntimePlugin<TPlugin, TInput, TError>
+    unsafe impl<TPlugin: 'static, TInput: ?Sized + 'static, TError: 'static, TMatch: ?Sized> Sync
+        for RuntimePlugin<TPlugin, TInput, TError, TMatch>
     {
     }
 
@@ -135,7 +140,8 @@ mod wasm_impls {
         TInput1: ?Sized + 'static,
         TInput2: ?Sized + 'static,
         TError: 'static,
-    > Send for RuntimePlugin2<TPlugin, TInput1, TInput2, TError>
+        TMatch: ?Sized,
+    > Send for RuntimePlugin2<TPlugin, TInput1, TInput2, TError, TMatch>
     {
     }
     unsafe impl<
@@ -143,7 +149,8 @@ mod wasm_impls {
         TInput1: ?Sized + 'static,
         TInput2: ?Sized + 'static,
         TError: 'static,
-    > Sync for RuntimePlugin2<TPlugin, TInput1, TInput2, TError>
+        TMatch: ?Sized,
+    > Sync for RuntimePlugin2<TPlugin, TInput1, TInput2, TError, TMatch>
     {
     }
 }